+++ resources/views/article/show.blade.php
@@ -23,21 +23,23 @@
 
   <hr>
 
-  <div class = "row">
+  @if($article->user_id === Auth::user()->id)
+    <div class = "row">
 
-    <div class = "col-md-1">
-      <a href="{{ route('article.edit', $article->id) }}"><button type="button" class="btn btn-danger">
-        <i class="fa fa-pencil fa-fw" aria-hidden="true"></i>Edit</button></a>
+      <div class = "col-md-1">
+        <a href="{{ route('article.edit', $article->id) }}"><button type="button" class="btn btn-danger">
+          <i class="fa fa-pencil fa-fw" aria-hidden="true"></i>Edit</button></a>
+      </div>
+
+      <div class = "col-md-1">
+        {!! Form::open(['method' => 'DELETE', 'route' => ['article.destroy', $article->id]]) !!}
+          <button type="submit" class="btn btn-danger">
+            <i class="fa fa-trash-o" aria-hidden="true"></i>Delete</button>
+        {!! Form::close() !!}
+      </div>
+
     </div>
-
-    <div class = "col-md-1">
-      {!! Form::open(['method' => 'DELETE', 'route' => ['article.destroy', $article->id]]) !!}
-        <button type="submit" class="btn btn-danger">
-          <i class="fa fa-trash-o" aria-hidden="true"></i>Delete</button>
-      {!! Form::close() !!}
-    </div>
-
-  </div>
+  @endif
 
 
 </div>
